fix: handle relation-broken in get_sidebar_items()
This adds the feature to get_sidebar_items to, when executed during a relation-broken event for this relation, omit the relation data for the breaking app (the app leaving the relation).  For context, see https://chat.charmhub.io/charmhub/pl/

Possibly controversially, this feature uses the Juju environment variables to determine what event type is being executed.  This is to avoid having to pass the `BoundEvent` object to every `get_sidebar_items()` call, and to keep this hack entirely in the relation library so the charm does not need to know about it.