build: install jsonschema from binary instead of building from src
Because some dependencies of jsonschema require the rust toolchain (see 1 and 2),
building the charm fails when building and installing some deps of jsonschema.
The workaround to this is to install jsonschema from binary, avoiding unnecesary
toolchains in the charm and reducing the chance of getting affected by third party
dependencies when they are updated.
[1]https://github.com/python-jsonschema/jsonschema/issues/1117
[2] https://github.com/python-jsonschema/jsonschema/issues/1114