build: install jsonschema from binary instead of building from src
Because some dependencies of jsonschema require the rust toolchain (see 1 and 2), building
the charm fails when building and installing some deps of jsonschema.
The workaround to this is to install jsonschema from binary, avoiding unnecesary toolchains
in the charm and reducing the chance of getting affected by third
party dependencies when they are updated.

Other considerations:
Other options were considered for fixing the issue, like installing rustc in the charm
directly and pinning jsonschema<4.18. The first option seem too invasive as charms will not
use rustc for anything during runtime. The last option restricted us from updating the package
in the event of a bug fix or if a new version is needed.
Installing from binary also has limitations, such as version management, which is now done
in charmcraft.yaml and not in the more common requirement files.

Fixes: #147

[1]python-jsonschema/jsonschema#1117
[2] python-jsonschema/jsonschema#1114