tests: refactor build and deploy test case
Remove the assertion that tests the kubeflow-dashboard-operator goes to Blocked
when the relation with kubeflow-profiles is not there. This assertion is already
covered in unit tests, and it is only adding noise at deployment time.
This commit ensures the dependencies like charms and relations are deployed on time
which will give time to everything to be setup before moving ahead with other test cases.